apps/files-service/pyproject.toml
[tool.poetry]
name = "files-service"
version = "0.1.0"
description = ""
authors = ["Cristian Alvarez H <alvarezhernandezcristian1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
boto3 = "^1.35.76"
python-ulid = "^3.0.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
opentelemetry-api = "^1.29.0"
opentelemetry-sdk = "^1.29.0"
opentelemetry-instrumentation-fastapi = "^0.50b0"
opentelemetry-exporter-otlp = "^1.29.0"
opentelemetry-instrumentation-boto = "^0.50b0"
opentelemetry-instrumentation-botocore = "^0.50b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
